feat: updated better health check
