+++ apps/api-languages/src/app/app.module.ts
@@ -11,8 +11,8 @@ import { DatadogTraceModule } from 'nestjs-ddtrace'
 import { LoggerModule } from 'nestjs-pino'
 
 import TranslationModule from '@core/nest/common/TranslationModule'
-import { NestHealthModule } from '@core/nest/health'
 
+import { NestHealthModule } from './modules/health/health.module'
 import { LanguageModule } from './modules/language/language.module'
 
 @Module({
