+++ apps/api-tags/src/app/app.module.ts
@@ -10,8 +10,8 @@ import { DatadogTraceModule } from 'nestjs-ddtrace'
 import { LoggerModule } from 'nestjs-pino'
 
 import TranslationModule from '@core/nest/common/TranslationModule'
-import { NestHealthModule } from '@core/nest/health'
 
+import { NestHealthModule } from './modules/health/health.module'
 import { TagModule } from './modules/tag/tag.module'
 
 @Module({
