+++ apps/api-videos/src/app/modules/health/health.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common'
+
+import { PrismaService } from '../../lib/prisma.service'
+
+import { NestHealthController } from './health.controller'
+
+@Module({
+  controllers: [NestHealthController],
+  providers: [PrismaService],
+  exports: []
+})
+export class NestHealthModule {}
