--- libs/nest/health/README.md
@@ -1,7 +1,0 @@
-# nest-health
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test nest-health` to execute the unit tests via [Jest](https://jestjs.io).
