--- libs/nest/health/jest.config.ts
@@ -1,18 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'nest-health',
-
-  globals: {},
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': [
-      'ts-jest',
-      {
-        tsconfig: '<rootDir>/tsconfig.spec.json'
-      }
-    ]
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../coverage/libs/nest/health',
-  preset: '../../../jest.preset.js'
-}
