--- libs/nest/health/project.json
@@ -1,23 +1,0 @@
-{
-  "name": "nest-health",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/nest/health/src",
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/nest/health/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/libs/nest/health"],
-      "options": {
-        "jestConfig": "libs/nest/health/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
