--- libs/nest/health/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/nest-health.module'
