--- libs/nest/health/src/lib/nest-health.controller.ts
@@ -1,9 +1,0 @@
-import { Controller, Get } from '@nestjs/common'
-
-@Controller()
-export class NestHealthController {
-  @Get('/.well-known/apollo/server-health')
-  getHealth(): string {
-    return 'OK'
-  }
-}
