--- libs/nest/health/src/lib/nest-health.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common'
-
-import { NestHealthController } from './nest-health.controller'
-
-@Module({
-  controllers: [NestHealthController],
-  providers: [],
-  exports: []
-})
-export class NestHealthModule {}
