+++ tsconfig.base.json
@@ -24,7 +24,6 @@
       "@core/nest/common/*": ["libs/nest/common/src/lib/*"],
       "@core/nest/decorators/*": ["libs/nest/decorators/src/lib/*"],
       "@core/nest/gqlAuthGuard/*": ["libs/nest/gqlAuthGuard/src/lib/*"],
-      "@core/nest/health": ["libs/nest/health/src/index.ts"],
       "@core/nest/powerBi/*": ["libs/nest/powerBi/src/lib/*"],
       "@core/shared/ui/*": [
         "libs/shared/ui/src/components/*",
